Add database fixture for tests
To enable automated testing for each section within the schema, each
test needs to attempt to insert data into a database running the latest
schema. This fixture installs the latest version of the schema into a
test collection and provides a base document and connection to the
newly created collection for any calling tests.