Restructure composite types to have a matrix of reference materials
To facilitate composite materials becoming a matrix of reference materials,
the list of available materials types and material properties are defined
as seperate items. A material object can now either be a reference material object,
or a composite material object where a matrix is declared of a base reference material
and a list of embedded reference materials is provided.