Coordinates support for mode shapes
Previously a mode shape within the eigen mode features
could only take an array a complex values; however, with
a mode shape, there needs to be support for the coordinates
of the sensors that are associated with value. As such the
mode shape property now has support for XYZ values for
coordinates and the readings.